fix(createWebWorkerPromise): Move up one directory
This results in:

-    worker = new Worker(new URL('../../web-workers/pipeline.worker.js', import.meta.url))
+    worker = new Worker(new URL('../web-workers/pipeline.worker.js', import.meta.url))

which helps the use case:

1) bundle with rollup
2) bundle the resulting bundle with vite
